requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
undetected-chromedriver>=3.5.3
selenium>=4.15.2
selenium-stealth>=1.0.6
python-dotenv>=1.0.0